Adjust calendar ranges: 1 month for index, 12 months for calendar page
As I'm mainly using this on my phone at the moment, I'm adjusting the
calendar range: show just the current month for the index page, and
show a whole year (but centered on the current month) for the calendar
page.

It would be nice to have this adjustable: either by pressing a button,
or by detecting screen size.

I may change the calendar page view to the past year, rather than 6
months before/after.

Signed-off-by: Hugh Brown (Saint Aardvark the Carpeted) <aardvark@saintaardvarkthecarpeted.com>